[SUS-76] notice까지 완료
