+++ src/notice/router/noticeItemRouter.ts
@@ -3,12 +3,30 @@ import multipartParser from '#util/multipartParser'
 import express from 'express'
 import NoticeEditBodyRequest from '../entity/dao/frontend/request/NoticeEditBodyRequest.js'
 import NoticePathRequest from '../entity/dao/frontend/request/NoticePathRequest.js'
+import WriteRequest from '../entity/dao/frontend/request/WriteRequest.js'
 import NoticeResponse from '../entity/dao/frontend/response/NoticeResponse.js'
+import WriteResponse from '../entity/dao/frontend/response/WriteResponse.js'
 import NoticeItemService from '../service/NoticeItemService.js'
 export const noticeItemRouter = express.Router()
 
+noticeItemRouter.post(
+  '/',
+  multipartParser('image', 5),
+  controller(
+    'admin',
+    null,
+    null,
+    WriteRequest,
+    WriteResponse,
+  )(async (req, res) => {
+    return res.send(
+      await NoticeItemService.writeNotice(req.memberIdx, req.body),
+    )
+  }),
+)
+
 noticeItemRouter.get(
-  '/:idx',
+  '/',
   controller(
     'login',
     null,
@@ -21,7 +39,7 @@ noticeItemRouter.get(
 )
 
 noticeItemRouter.delete(
-  '/:idx',
+  '/',
   controller(
     'admin',
     null,
@@ -35,7 +53,7 @@ noticeItemRouter.delete(
 )
 
 noticeItemRouter.patch(
-  '/:idx',
+  '/',
   multipartParser('image', 5),
   controller(
     'admin',
