+++ src/notice/service/ListService.ts
@@ -1,44 +1,29 @@
 import ListRequest from '../entity/dao/frontend/request/ListRequest.js'
-import ListSearchRequest from '../entity/dao/frontend/request/ListSearchRequest.js'
 import ListResponse from '../entity/dao/frontend/response/ListResponse.js'
-import ListSearchResponse from '../entity/dao/frontend/response/ListSearchResponse.js'
-import NoticeList from '../entity/dto/NoticeList.js'
 import NoticeRepository from '../repository/NoticeRepository.js'
 
 export default class ListService {
   static async getList(listRequest: ListRequest) {
-    const { current, display } = listRequest
+    const { current, display, title } = listRequest
     const offset = (current - 1) * display
-    const listData: any = await NoticeRepository.getPagedListFromDb(
-      display,
-      offset,
-    )
 
-    const noticeList = new NoticeList(
+    let listData
+    if (title) {
+      const titleToSearch = '%' + title + '%'
+      listData = await NoticeRepository.getSearchListFromDb(
+        titleToSearch,
+        display,
+        offset,
+      )
+    } else {
+      listData = await NoticeRepository.getPagedListFromDb(display, offset)
+    }
+
+    return new ListResponse(
       current,
       display,
       listData.totalCountResult,
       listData.listResult,
     )
-
-    return ListResponse.of(noticeList)
-  }
-
-  static async searchList(listSearchRequest: ListSearchRequest) {
-    const { title, display, current } = listSearchRequest
-    const offset = (current - 1) * display
-    const titleToSearch = '%' + title + '%'
-    const searchData: any = await NoticeRepository.getSearchListFromDb(
-      titleToSearch,
-      display,
-      offset,
-    )
-    const noticeList = new NoticeList(
-      current,
-      display,
-      searchData.totalCountResult,
-      searchData.listResult,
-    )
-    return ListSearchResponse.of(noticeList)
   }
 }
